arch: split agent.rs into registry, process_builder, agent_process (#280)
Split the 2152-line agent.rs god module into three focused submodules:
- agent_registry.rs: CRUD, state persistence, one-shot send/spawn
- process_builder.rs: command construction (pure functions)
- agent_process.rs: AgentProcess persistent lifecycle + Executor impl

agent.rs becomes a thin barrel module re-exporting everything, so all
existing callers (worker, acp, serve, commands, mcp) compile unchanged.
