+++ src/app/agent.rs
@@ -1,1939 +1,26 @@
-use anyhow::{Context, Result, bail};
-use chrono::Utc;
-use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
-use std::process::Stdio;
-use tokio::io::{AsyncBufReadExt, AsyncReadExt, AsyncWriteExt};
-use tokio::process::Command;
-use tracing::{debug, error, info, warn};
-
-use crate::config::{self, ContainerConfig, UserConfig};
-use crate::infra::dto::{ConfigAgentRuntime, ConfigContextConfig, ConfigSessionMode};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentConfig {
-    pub name: String,
-    pub model: String,
-    pub system_prompt: String,
-    pub work_dir: String,
-    pub max_turns: u32,
-    /// Optional Linux user to run the agent process as.
-    #[serde(default)]
-    pub unix_user: Option<String>,
-    /// Budget cap in USD.
-    #[serde(default = "default_budget_usd")]
-    pub budget_usd: f64,
-    /// Command to run. Defaults to ["claude"].
-    #[serde(default = "default_agent_command")]
-    pub command: Vec<String>,
-    /// Path to the agent's deskd.yaml (injected as DESKD_AGENT_CONFIG into claude).
-    #[serde(default)]
-    pub config_path: Option<String>,
-    /// Container config. When set, the agent runs inside a container.
-    #[serde(default)]
-    pub container: Option<ContainerConfig>,
-    /// Session mode: persistent (default) or ephemeral.
-    #[serde(default)]
-    pub session: ConfigSessionMode,
-    /// Agent runtime protocol: claude (default) or acp.
-    #[serde(default)]
-    pub runtime: ConfigAgentRuntime,
-    /// Context system configuration (main branch, compaction).
-    #[serde(default)]
-    pub context: Option<ConfigContextConfig>,
-}
-
-fn default_budget_usd() -> f64 {
-    50.0
-}
-
-fn default_agent_command() -> Vec<String> {
-    vec!["claude".to_string()]
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentState {
-    pub config: AgentConfig,
-    pub pid: u32,
-    pub session_id: String,
-    pub total_turns: u32,
-    pub total_cost: f64,
-    pub created_at: String,
-    /// "idle" or "working"
-    #[serde(default = "default_status")]
-    pub status: String,
-    /// Truncated text of the task currently being processed.
-    #[serde(default)]
-    pub current_task: String,
-    /// Name of the parent agent that spawned this sub-agent, if any.
-    #[serde(default)]
-    pub parent: Option<String>,
-}
-
-fn default_status() -> String {
-    "idle".to_string()
-}
-
-fn state_path(name: &str) -> PathBuf {
-    config::state_dir().join(format!("{}.yaml", name))
-}
-
-fn log_path(name: &str) -> PathBuf {
-    config::log_dir().join(format!("{}.log", name))
-}
-
-/// Path to the agent's stderr log file.
-pub fn stderr_log_path(name: &str) -> PathBuf {
-    config::log_dir().join(format!("{}.stderr.log", name))
-}
-
-/// Path to the agent's stream-json log file.
-pub fn stream_log_path(name: &str) -> PathBuf {
-    let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".into());
-    let dir = PathBuf::from(home).join(".deskd").join("logs").join(name);
-    dir.join("stream.jsonl")
-}
-
-/// Rotate stream log: if file exceeds max_bytes, keep the last half.
-fn rotate_stream_log(path: &PathBuf, max_bytes: u64) {
-    let Ok(meta) = std::fs::metadata(path) else {
-        return;
-    };
-    if meta.len() <= max_bytes {
-        return;
-    }
-    let Ok(content) = std::fs::read_to_string(path) else {
-        return;
-    };
-    // Keep the last half of lines.
-    let lines: Vec<&str> = content.lines().collect();
-    let keep_from = lines.len() / 2;
-    let truncated: String = lines[keep_from..].iter().flat_map(|l| [*l, "\n"]).collect();
-    let _ = std::fs::write(path, truncated);
-}
-
-pub fn load_state(name: &str) -> Result<AgentState> {
-    let path = state_path(name);
-    let content =
-        std::fs::read_to_string(&path).with_context(|| format!("Agent '{}' not found", name))?;
-    let state: AgentState = serde_yaml::from_str(&content)?;
-    Ok(state)
-}
-
-fn save_state(state: &AgentState) -> Result<()> {
-    let path = state_path(&state.config.name);
-    let content = serde_yaml::to_string(state)?;
-    std::fs::write(&path, content)?;
-    Ok(())
-}
-
-pub fn save_state_pub(state: &AgentState) -> Result<()> {
-    save_state(state)
-}
-
-/// Create a new agent (saves state file; does not start a worker process).
-pub async fn create(cfg: &AgentConfig) -> Result<AgentState> {
-    if state_path(&cfg.name).exists() {
-        bail!("Agent '{}' already exists. Remove it first.", cfg.name);
-    }
-
-    let state = AgentState {
-        config: cfg.clone(),
-        pid: 0,
-        session_id: String::new(),
-        total_turns: 0,
-        total_cost: 0.0,
-        created_at: Utc::now().to_rfc3339(),
-        status: "idle".to_string(),
-        current_task: String::new(),
-        parent: None,
-    };
-
-    save_state(&state)?;
-    info!(agent = %cfg.name, "agent created");
-    Ok(state)
-}
-
-/// Create or update agent state from an AgentConfig.
-/// If state already exists, updates the config fields but preserves session/cost/turns.
-/// Used for sub-agents spawned from deskd.yaml.
-pub async fn create_or_update_from_config(cfg: &AgentConfig) -> Result<AgentState> {
-    let path = state_path(&cfg.name);
-    if path.exists() {
-        let mut state = load_state(&cfg.name)?;
-        state.config = cfg.clone();
-        save_state(&state)?;
-        info!(agent = %cfg.name, "sub-agent state updated");
-        return Ok(state);
-    }
-    let state = AgentState {
-        config: cfg.clone(),
-        pid: 0,
-        session_id: String::new(),
-        total_turns: 0,
-        total_cost: 0.0,
-        created_at: Utc::now().to_rfc3339(),
-        status: "idle".to_string(),
-        current_task: String::new(),
-        parent: None,
-    };
-    save_state(&state)?;
-    info!(agent = %cfg.name, "sub-agent created");
-    Ok(state)
-}
-
-/// Create or recover agent state from workspace AgentDef + optional UserConfig.
-/// If state already exists, returns it with config fields updated from current def.
-/// Model priority: workspace def.model override > user_cfg.model > default.
-pub async fn create_or_recover(
-    def: &config::AgentDef,
-    user_cfg: Option<&UserConfig>,
-) -> Result<AgentState> {
-    let model = def
-        .model
-        .clone()
-        .or_else(|| user_cfg.map(|c| c.model.clone()))
-        .unwrap_or_else(|| "claude-sonnet-4-6".to_string());
-
-    let system_prompt = user_cfg
-        .map(|c| c.system_prompt.clone())
-        .unwrap_or_default();
-
-    let max_turns = user_cfg.map(|c| c.max_turns).unwrap_or(100);
-
-    let cfg = AgentConfig {
-        name: def.name.clone(),
-        model,
-        system_prompt,
-        work_dir: def.work_dir.clone(),
-        max_turns,
-        unix_user: def.unix_user.clone(),
-        budget_usd: def.budget_usd,
-        command: def.command.clone(),
-        config_path: Some(def.config_path()),
-        container: def.container.clone(),
-        session: ConfigSessionMode::default(),
-        runtime: def.runtime.clone(),
-        context: user_cfg.and_then(|c| c.context.clone()),
-    };
-
-    let path = state_path(&def.name);
-    if path.exists() {
-        let mut state = load_state(&def.name)?;
-        // Update mutable fields on recovery; preserve session_id + costs.
-        state.config = cfg;
-        save_state(&state)?;
-        info!(agent = %def.name, session_id = %state.session_id, "recovered existing agent state");
-        return Ok(state);
-    }
-
-    let state = AgentState {
-        config: cfg,
-        pid: 0,
-        session_id: String::new(),
-        total_turns: 0,
-        total_cost: 0.0,
-        created_at: Utc::now().to_rfc3339(),
-        status: "idle".to_string(),
-        current_task: String::new(),
-        parent: None,
-    };
-    save_state(&state)?;
-    info!(agent = %def.name, "agent created");
-    Ok(state)
-}
-
-/// Run claude for one task, return the response text.
-///
-/// `bus_socket`: path to the agent's bus (injected as DESKD_BUS_SOCKET).
-/// Claude uses this to call the `send_message` MCP tool.
-pub async fn send(
-    name: &str,
-    message: &str,
-    max_turns: Option<u32>,
-    bus_socket: Option<&str>,
-) -> Result<String> {
-    send_inner(name, message, max_turns, bus_socket, None, None, None).await
-}
-
-/// Format a plain text message as a stream-json user turn line (with trailing newline).
-fn format_user_message(text: &str) -> String {
-    let msg = serde_json::json!({
-        "type": "user",
-        "message": {
-            "role": "user",
-            "content": [{"type": "text", "text": text}]
-        }
-    });
-    let mut line = serde_json::to_string(&msg).expect("json serialization cannot fail");
-    line.push('\n');
-    line
-}
-
-async fn send_inner(
-    name: &str,
-    message: &str,
-    max_turns: Option<u32>,
-    bus_socket: Option<&str>,
-    progress_tx: Option<tokio::sync::mpsc::UnboundedSender<String>>,
-    image: Option<(&str, &str)>,
-    inject_rx: Option<tokio::sync::mpsc::UnboundedReceiver<String>>,
-) -> Result<String> {
-    let mut state = load_state(name)?;
-
-    let turns = max_turns.unwrap_or(state.config.max_turns);
-
-    // Always use stream-json input so we can keep stdin open for mid-task injection.
-    let mut args: Vec<String> = Vec::new();
-
-    // Use --resume only when session is persistent and session_id exists.
-    let use_resume =
-        state.config.session == ConfigSessionMode::Persistent && !state.session_id.is_empty();
-
-    if use_resume {
-        args.push("--resume".to_string());
-        args.push(state.session_id.clone());
-    }
-
-    if !state.config.system_prompt.is_empty() && !use_resume {
-        args.push("--system-prompt".to_string());
-        args.push(state.config.system_prompt.clone());
-    }
-
-    // Auto-inject required flags for stream-json operation (#151).
-    inject_required_flags(&mut args, &state.config.command);
-
-    // Auto-inject --mcp-config for deskd tools when not already present (#153).
-    if !state
-        .config
-        .command
-        .iter()
-        .any(|a| a.contains("mcp-config"))
-    {
-        let deskd_bin = std::env::var("DESKD_BIN").unwrap_or_else(|_| "deskd".to_string());
-        let mcp_json = serde_json::json!({
-            "mcpServers": {
-                "deskd": {
-                    "command": deskd_bin,
-                    "args": ["mcp", "--agent", name]
-                }
-            }
-        })
-        .to_string();
-        args.push("--mcp-config".to_string());
-        args.push(mcp_json);
-    }
-
-    // Inject --model from agent config (sourced from deskd.yaml) unless the command
-    // array already contains --model (e.g. hardcoded in workspace.yaml).
-    if !state.config.model.is_empty()
-        && !state
-            .config
-            .command
-            .iter()
-            .any(|a| a == "--model" || a.starts_with("--model="))
-    {
-        args.push("--model".to_string());
-        args.push(state.config.model.clone());
-    }
-
-    debug!(agent = %name, turns, model = %state.config.model, multimodal = image.is_some(), "spawning claude");
-
-    // Env vars injected into the claude process:
-    //   DESKD_BUS_SOCKET  — bus socket for MCP send_message tool
-    //   DESKD_AGENT_NAME  — agent name for MCP server
-    //   DESKD_AGENT_CONFIG — path to deskd.yaml for tool description generation
-    let bus_path = bus_socket
-        .map(|s| s.to_string())
-        .unwrap_or_else(|| config::agent_bus_socket(&state.config.work_dir));
-
-    let config_path_str = state.config.config_path.clone().unwrap_or_default();
-    let mut extra_env: Vec<(&str, &str)> =
-        vec![("DESKD_AGENT_NAME", name), ("DESKD_BUS_SOCKET", &bus_path)];
-    if !config_path_str.is_empty() {
-        extra_env.push(("DESKD_AGENT_CONFIG", &config_path_str));
-    }
-
-    let mut cmd = build_command(&state.config, &args, &extra_env);
-    cmd.stdin(Stdio::piped());
-    let mut child = cmd.spawn().context("Failed to spawn claude CLI")?;
-
-    // Internal channel: all lines that should be written to Claude's stdin flow through here.
-    let (stdin_line_tx, mut stdin_line_rx) = tokio::sync::mpsc::unbounded_channel::<String>();
-
-    // Build and send the initial user message (text + optional image).
-    let initial_msg = if let Some((b64_data, media_type)) = image {
-        let msg = serde_json::json!({
-            "type": "user",
-            "message": {
-                "role": "user",
-                "content": [
-                    {
-                        "type": "image",
-                        "source": {
-                            "type": "base64",
-                            "media_type": media_type,
-                            "data": b64_data,
-                        }
-                    },
-                    {"type": "text", "text": message}
-                ]
-            }
-        });
-        let mut line = serde_json::to_string(&msg)?;
-        line.push('\n');
-        line
-    } else {
-        format_user_message(message)
-    };
-
-    stdin_line_tx
-        .send(initial_msg)
-        .expect("channel is open at this point");
-
-    // If inject_rx is provided, spawn a task to forward injected messages as new user turns.
-    if let Some(mut rx) = inject_rx {
-        let tx = stdin_line_tx.clone();
-        tokio::spawn(async move {
-            while let Some(msg) = rx.recv().await {
-                let line = format_user_message(&msg);
-                if tx.send(line).is_err() {
-                    break;
-                }
-            }
-        });
-    }
-
-    // Stdin writer task: reads lines from the internal channel and writes them to Claude's stdin.
-    // When all senders are dropped (inject task exits + we drop stdin_line_tx below),
-    // stdin_line_rx drains and the task ends, closing Claude's stdin.
-    let mut child_stdin = child.stdin.take().expect("stdin is piped");
-    tokio::spawn(async move {
-        while let Some(line) = stdin_line_rx.recv().await {
-            if child_stdin.write_all(line.as_bytes()).await.is_err() {
-                break;
-            }
-        }
-        // stdin closes here when all senders are dropped
-    });
-
-    // Drop our copy of the sender so that when the inject forwarder also exits,
-    // the writer task sees EOF and closes Claude's stdin naturally.
-    drop(stdin_line_tx);
-
-    let stdout = child.stdout.take().expect("stdout is piped");
-    let stderr = child.stderr.take().expect("stderr is piped");
-
-    // Read stderr in background so it doesn't block stdout parsing.
-    let stderr_task = tokio::spawn(async move {
-        let mut buf = String::new();
-        let mut reader = tokio::io::BufReader::new(stderr);
-        let _ = reader.read_to_string(&mut buf).await;
-        buf
-    });
-
-    let mut lines = tokio::io::BufReader::new(stdout).lines();
-    let mut response_text = String::new();
-    let mut new_session_id = String::new();
-    let mut task_cost = 0.0;
-    let mut task_turns = 0u32;
-
-    while let Some(line) = lines.next_line().await? {
-        if let Ok(v) = serde_json::from_str::<serde_json::Value>(&line) {
-            match v.get("type").and_then(|t| t.as_str()) {
-                Some("assistant") => {
-                    let mut block_text = String::new();
-                    if let Some(blocks) = v
-                        .get("message")
-                        .and_then(|m| m.get("content"))
-                        .and_then(|c| c.as_array())
-                    {
-                        for block in blocks {
-                            if block.get("type").and_then(|t| t.as_str()) == Some("text")
-                                && let Some(text) = block.get("text").and_then(|t| t.as_str())
-                            {
-                                block_text.push_str(text);
-                            }
-                        }
-                    }
-                    if !block_text.is_empty() {
-                        if let Some(tx) = &progress_tx {
-                            let _ = tx.send(block_text.clone());
-                        }
-                        response_text.push_str(&block_text);
-                    }
-                }
-                Some("result") => {
-                    if let Some(sid) = v.get("session_id").and_then(|s| s.as_str()) {
-                        new_session_id = sid.to_string();
-                    }
-                    if let Some(cost) = v.get("total_cost_usd").and_then(|c| c.as_f64()) {
-                        task_cost = cost;
-                    }
-                    if let Some(t) = v.get("num_turns").and_then(|t| t.as_u64()) {
-                        task_turns = t as u32;
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    let _ = child.wait().await;
-    let stderr_str = stderr_task.await.unwrap_or_default();
-
-    // Only save session_id for persistent agents.
-    if state.config.session == ConfigSessionMode::Persistent && !new_session_id.is_empty() {
-        state.session_id = new_session_id;
-    }
-    state.total_cost += task_cost;
-    state.total_turns += task_turns;
-    save_state(&state)?;
-
-    if response_text.is_empty() {
-        if !stderr_str.is_empty() {
-            bail!("Claude error: {}", stderr_str.trim());
-        }
-        // If we used --resume and got no response, the session_id is stale.
-        // Clear it and retry once with a fresh session.
-        if !state.session_id.is_empty() {
-            warn!(agent = %name, session_id = %state.session_id, "stale session_id — retrying without --resume");
-            state.session_id = String::new();
-            save_state(&state)?;
-            return Box::pin(send_inner(
-                name,
-                message,
-                max_turns,
-                bus_socket,
-                progress_tx,
-                image,
-                None, // inject_rx consumed; don't retry injection
-            ))
-            .await;
-        }
-        bail!("No response from claude");
-    }
-
-    Ok(response_text)
-}
-
-/// Build the tokio Command for running the agent process.
-/// Inject flags required for persistent stream-json operation.
-///
-/// Checks the agent's command array to avoid duplicating flags that were
-/// explicitly set in workspace.yaml. Sub-agents created via MCP typically
-/// have `command: ["claude"]` with no flags, so all are injected.
-fn inject_required_flags(args: &mut Vec<String>, command: &[String]) {
-    // --input-format=stream-json — needed for multimodal content and mid-task injection.
-    args.push("--input-format=stream-json".to_string());
-
-    // --output-format stream-json — worker expects JSON stdout, not plain text.
-    if !command.iter().any(|a| a.contains("output-format")) {
-        args.push("--output-format".to_string());
-        args.push("stream-json".to_string());
-    }
-
-    // --verbose — required when using --output-format stream-json.
-    if !command.iter().any(|a| a == "--verbose") {
-        args.push("--verbose".to_string());
-    }
-
-    // --dangerously-skip-permissions — without this, Claude blocks on permission prompts.
-    if !command
-        .iter()
-        .any(|a| a.contains("dangerously-skip-permissions"))
-    {
-        args.push("--dangerously-skip-permissions".to_string());
-    }
-}
-
-/// Uses cfg.command as the executable (defaults to ["claude"]).
-/// When unix_user is set, wraps with sudo and strips SSH env vars.
-/// extra_env: additional environment variables to pass to the process.
-pub fn build_command(cfg: &AgentConfig, args: &[String], extra_env: &[(&str, &str)]) -> Command {
-    if let Some(ref container) = cfg.container {
-        return build_container_command(cfg, container, args, extra_env);
-    }
-
-    let (bin, prefix) = split_command(&cfg.command);
-    let mut cmd = match &cfg.unix_user {
-        Some(user) => {
-            let mut c = Command::new("sudo");
-            c.args(["-u", user, "-H", "--", "env"]);
-            for (k, v) in extra_env {
-                c.arg(format!("{}={}", k, v));
-            }
-            c.arg(bin);
-            c.args(prefix);
-            c.args(args);
-            c.env_remove("SSH_AUTH_SOCK");
-            c.env_remove("SSH_AGENT_PID");
-            c
-        }
-        None => {
-            let mut c = Command::new(bin);
-            c.args(prefix);
-            c.args(args);
-            c
-        }
-    };
-    if cfg.unix_user.is_none() {
-        for (k, v) in extra_env {
-            cmd.env(k, v);
-        }
-    }
-    cmd.current_dir(&cfg.work_dir)
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped());
-    cmd
-}
-
-/// Build a command that runs the agent process inside a container.
-///
-/// Constructs: `<runtime> run --rm -i --name deskd-<agent>
-///   -v <work_dir>:<work_dir> -w <work_dir>
-///   -v <bus_socket_dir>:<bus_socket_dir>
-///   [-v <mount>...] [-v <volume>...] [-e <key>=<val>...]
-///   <image> <command> <args>`
-fn build_container_command(
-    cfg: &AgentConfig,
-    container: &ContainerConfig,
-    args: &[String],
-    extra_env: &[(&str, &str)],
-) -> Command {
-    let runtime = &container.runtime;
-    let mut cmd = Command::new(runtime);
-
-    cmd.args(["run", "--rm", "-i"]);
-    cmd.args(["--name", &format!("deskd-{}", cfg.name)]);
-
-    // Mount the agent's work_dir so the process can access the repo.
-    cmd.args(["-v", &format!("{}:{}", cfg.work_dir, cfg.work_dir)]);
-    cmd.args(["-w", &cfg.work_dir]);
-
-    // The bus socket is at {work_dir}/.deskd/bus.sock — already covered by work_dir mount.
-
-    // User-defined mounts from container config.
-    for mount in &container.mounts {
-        let expanded = normalize_mount(&expand_tilde(mount));
-        cmd.args(["-v", &expanded]);
-    }
-
-    // Docker volumes.
-    for vol in &container.volumes {
-        cmd.args(["-v", vol]);
-    }
-
-    // Environment variables from container config.
-    for (k, v) in &container.env {
-        cmd.args(["-e", &format!("{}={}", k, v)]);
-    }
-
-    // Extra env vars (DESKD_BUS_SOCKET, DESKD_AGENT_NAME, DESKD_AGENT_CONFIG).
-    for (k, v) in extra_env {
-        cmd.args(["-e", &format!("{}={}", k, v)]);
-    }
-
-    // If there's a config_path, mount it into the container.
-    if let Some(ref config_path) = cfg.config_path {
-        let cp = PathBuf::from(config_path);
-        if let Some(parent) = cp.parent() {
-            let parent_str = parent.to_string_lossy();
-            // Only mount if not already under work_dir.
-            if !parent_str.starts_with(&cfg.work_dir) {
-                cmd.args(["-v", &format!("{}:{}", parent_str, parent_str)]);
-            }
-        }
-    }
-
-    // Image.
-    cmd.arg(&container.image);
-
-    // The actual command to run inside the container.
-    let (bin, prefix) = split_command(&cfg.command);
-    cmd.arg(bin);
-    cmd.args(prefix);
-    cmd.args(args);
-
-    // Don't set current_dir on the host — the container uses -w.
-    cmd.stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped());
-    cmd
-}
-
-/// Expand ~ to $HOME in mount paths.
-/// Handles formats: "~/foo", "~/foo:ro", "~/foo:~/bar", "~/foo:~/bar:ro".
-fn expand_tilde(path: &str) -> String {
-    let home = match std::env::var("HOME") {
-        Ok(h) => h,
-        Err(_) => return path.to_string(),
-    };
-
-    // Split into parts, handling the colon-separated mount syntax.
-    // Possible formats: "src", "src:ro", "src:dst", "src:dst:ro"
-    let parts: Vec<&str> = path.splitn(3, ':').collect();
-    let expand = |s: &str| -> String {
-        if s.starts_with("~/") || s == "~" {
-            s.replacen('~', &home, 1)
-        } else {
-            s.to_string()
-        }
-    };
-
-    match parts.len() {
-        1 => expand(parts[0]),
-        2 => format!("{}:{}", expand(parts[0]), expand(parts[1])),
-        3 => format!("{}:{}:{}", expand(parts[0]), expand(parts[1]), parts[2]),
-        _ => path.to_string(),
-    }
-}
-
-/// Normalize a mount spec to always have src:dst[:opts].
-/// "path" → "path:path"
-/// "path:ro" → "path:path:ro"
-/// "src:dst" → "src:dst" (unchanged)
-/// "src:dst:ro" → "src:dst:ro" (unchanged)
-fn normalize_mount(mount: &str) -> String {
-    let parts: Vec<&str> = mount.splitn(3, ':').collect();
-    match parts.len() {
-        1 => {
-            // "path" → "path:path"
-            format!("{}:{}", parts[0], parts[0])
-        }
-        2 => {
-            // Could be "src:dst" or "path:ro"/"path:rw"
-            if parts[1] == "ro" || parts[1] == "rw" {
-                format!("{}:{}:{}", parts[0], parts[0], parts[1])
-            } else {
-                mount.to_string()
-            }
-        }
-        _ => mount.to_string(),
-    }
-}
-
-fn split_command(command: &[String]) -> (&str, &[String]) {
-    match command {
-        [] => ("claude", &[]),
-        [bin] => (bin.as_str(), &[]),
-        [bin, rest @ ..] => (bin.as_str(), rest),
-    }
-}
-
-/// Spawn an ephemeral sub-agent: create, run task, clean up.
-/// Returns the agent's text response.
-pub async fn spawn_ephemeral(
-    name: &str,
-    task: &str,
-    model: &str,
-    work_dir: &str,
-    max_turns: u32,
-    bus_socket: &str,
-    parent_name: &str,
-) -> Result<String> {
-    let unique_name = format!("{}-{}", name, uuid::Uuid::new_v4().as_simple());
-
-    let cfg = AgentConfig {
-        name: unique_name.clone(),
-        model: model.to_string(),
-        system_prompt: String::new(),
-        work_dir: work_dir.to_string(),
-        max_turns,
-        unix_user: None,
-        budget_usd: 50.0,
-        command: default_agent_command(),
-        config_path: None,
-        container: None,
-        session: ConfigSessionMode::default(),
-        runtime: ConfigAgentRuntime::default(),
-        context: None,
-    };
-
-    create(&cfg).await?;
-    info!(agent = %unique_name, parent = %parent_name, bus = %bus_socket, "spawning ephemeral sub-agent");
-
-    let result = send(&unique_name, task, Some(max_turns), Some(bus_socket)).await;
-
-    if let Err(e) = remove(&unique_name).await {
-        warn!(agent = %unique_name, error = %e, "failed to clean up ephemeral agent");
-    }
-
-    result
-}
-
-/// List all agents whose state files exist on disk.
-pub async fn list() -> Result<Vec<AgentState>> {
-    let dir = config::state_dir();
-    let mut agents = Vec::new();
-
-    if let Ok(entries) = std::fs::read_dir(&dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.extension().map(|e| e == "yaml").unwrap_or(false)
-                && let Ok(content) = std::fs::read_to_string(&path)
-                && let Ok(state) = serde_yaml::from_str::<AgentState>(&content)
-            {
-                agents.push(state);
-            }
-        }
-    }
-
-    Ok(agents)
-}
-
-/// Build a domain `Agent` from an `AgentState`, checking process health.
-///
-/// `live_agents` is the set of agent names currently registered on a bus.
-/// If the agent's PID is gone and it's not in `live_agents`, it's marked Unhealthy.
-pub fn to_domain_agent(
-    state: &AgentState,
-    live_agents: &std::collections::HashSet<String>,
-) -> crate::domain::agent::Agent {
-    use crate::domain::agent::{Agent, AgentCapabilities, AgentRuntime, AgentStatus, SessionMode};
-
-    let runtime = match state.config.runtime {
-        ConfigAgentRuntime::Claude => AgentRuntime::Claude,
-        ConfigAgentRuntime::Acp => AgentRuntime::Acp,
-    };
-
-    let session_mode = match state.config.session {
-        ConfigSessionMode::Persistent => SessionMode::Persistent,
-        ConfigSessionMode::Ephemeral => SessionMode::Ephemeral,
-    };
-
-    let status = if state.status == "working" {
-        AgentStatus::Busy {
-            task_id: state.current_task.clone(),
-        }
-    } else if live_agents.contains(&state.config.name) {
-        AgentStatus::Ready
-    } else if state.pid > 0 && std::path::Path::new(&format!("/proc/{}", state.pid)).exists() {
-        // Process alive but not yet registered on bus — treat as Ready.
-        AgentStatus::Ready
-    } else if state.pid > 0 {
-        AgentStatus::Unhealthy {
-            since: chrono::Utc::now().to_rfc3339(),
-            reason: format!("process {} not found", state.pid),
-        }
-    } else {
-        // pid == 0 means agent was never started or was cleaned up.
-        AgentStatus::Ready
-    };
-
-    Agent {
-        name: state.config.name.clone(),
-        runtime,
-        session_mode,
-        capabilities: AgentCapabilities {
-            model: state.config.model.clone(),
-            labels: Vec::new(),
-        },
-        status,
-    }
-}
-
-/// Remove an agent (state file + log).
-pub async fn remove(name: &str) -> Result<()> {
-    let path = state_path(name);
-    if !path.exists() {
-        bail!("Agent '{}' not found", name);
-    }
-    std::fs::remove_file(&path)?;
-    let log = log_path(name);
-    if log.exists()
-        && let Err(e) = std::fs::remove_file(&log)
-    {
-        warn!(agent = %name, error = %e, "failed to remove log file");
-    }
-    let stderr_log = stderr_log_path(name);
-    if stderr_log.exists()
-        && let Err(e) = std::fs::remove_file(&stderr_log)
-    {
-        warn!(agent = %name, error = %e, "failed to remove stderr log file");
-    }
-    let stream_log = stream_log_path(name);
-    if stream_log.exists()
-        && let Err(e) = std::fs::remove_file(&stream_log)
-    {
-        warn!(agent = %name, error = %e, "failed to remove stream log file");
-    }
-    info!(agent = %name, "agent removed");
-    Ok(())
-}
-
-// ─── Persistent agent process ─────────────────────────────────────────────────
+//! Agent module — re-exports from submodules for backward compatibility.
+//!
+//! Split into three submodules (#280):
+//! - `agent_registry` — CRUD, state persistence, one-shot execution
+//! - `process_builder` — command construction (pure functions)
+//! - `agent_process` — persistent process lifecycle (`AgentProcess` + `Executor`)
+
+// Re-export everything so `crate::app::agent::*` continues to work.
+pub use super::agent_process::AgentProcess;
+pub use super::agent_registry::{
+    AgentConfig, AgentState, create, create_or_recover, create_or_update_from_config,
+    default_agent_command, list, load_state, remove, save_state_pub, send, spawn_ephemeral,
+    stderr_log_path, stream_log_path, to_domain_agent,
+};
+pub use super::process_builder::build_command;
 
 // Re-export executor port types — canonical definitions live in ports::executor.
 pub use crate::ports::executor::{Executor, TaskLimits, TokenUsage, TurnResult};
 
-/// Events emitted by the stdout reader task.
-enum StdoutEvent {
-    /// A text block from an `assistant` message, with optional usage data.
-    TextBlock(String, Option<TokenUsage>),
-    /// The `result` event marking end of a turn.
-    Result(TurnResult),
-    /// Process exited (stdout closed) with diagnostic context.
-    ProcessExited {
-        exit_code: Option<i32>,
-        stderr_tail: String,
-        lifetime_secs: u64,
-        used_resume: bool,
-    },
-}
-
-/// Truncate stderr output for display in error messages (max 200 chars).
-fn truncate_stderr(s: &str) -> &str {
-    let trimmed = s.trim();
-    if trimmed.len() <= 200 {
-        trimmed
-    } else {
-        &trimmed[trimmed.len() - 200..]
-    }
-}
-
-/// Check whether a freshly spawned process stays alive for at least 5 seconds.
-///
-/// Returns `Ok(event_rx)` if the process is still running after the window.
-/// Returns `Err(...)` with a descriptive message if the process exits immediately —
-/// indicating a bad image, bad mounts, auth failure, or similar startup error.
-async fn check_fresh_startup(
-    name: &str,
-    mut event_rx: tokio::sync::mpsc::UnboundedReceiver<StdoutEvent>,
-) -> Result<tokio::sync::mpsc::UnboundedReceiver<StdoutEvent>> {
-    let result = tokio::time::timeout(std::time::Duration::from_secs(5), async {
-        loop {
-            match event_rx.try_recv() {
-                Ok(StdoutEvent::ProcessExited {
-                    exit_code,
-                    stderr_tail,
-                    ..
-                }) => return Some((exit_code, stderr_tail)),
-                Ok(_) => {
-                    // Got a real event — process is alive.
-                    return None;
-                }
-                Err(tokio::sync::mpsc::error::TryRecvError::Empty) => {
-                    tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-                }
-                Err(_) => {
-                    // Channel closed unexpectedly.
-                    return Some((None, String::new()));
-                }
-            }
-        }
-    })
-    .await
-    .unwrap_or(None); // timeout reached = process is alive
-
-    if let Some((exit_code, stderr_tail)) = result {
-        let code_str = exit_code
-            .map(|c| c.to_string())
-            .unwrap_or_else(|| "?".to_string());
-        error!(
-            agent = %name,
-            exit_code = %code_str,
-            stderr = %truncate_stderr(&stderr_tail),
-            "container/process exited immediately after fresh start — check image, mounts, and auth"
-        );
-        bail!(
-            "process for agent '{}' exited immediately (exit_code={}). stderr: {}",
-            name,
-            code_str,
-            truncate_stderr(&stderr_tail)
-        );
-    }
-
-    Ok(event_rx)
-}
-
-/// A long-lived Claude process that accepts multiple tasks via stdin.
-///
-/// Usage:
-///   let process = AgentProcess::start(name, bus_socket).await?;
-///   let result = process.send_task(message, progress_tx, None, None).await?;
-///   // process stays alive for the next task
-///   process.stop().await;
-pub struct AgentProcess {
-    /// Send lines to Claude's stdin.
-    stdin_tx: tokio::sync::mpsc::UnboundedSender<String>,
-    /// Receive stdout events (text blocks + result).
-    event_rx: tokio::sync::Mutex<tokio::sync::mpsc::UnboundedReceiver<StdoutEvent>>,
-    /// Child process handle for shutdown (shared with stdout reader for exit code capture).
-    child: std::sync::Arc<tokio::sync::Mutex<Option<tokio::process::Child>>>,
-    /// Agent name.
-    name: String,
-    /// Last cumulative cost reported by Claude (for computing deltas).
-    /// Claude's `total_cost_usd` is session-cumulative, not per-task.
-    last_reported_cost: tokio::sync::Mutex<f64>,
-    /// Last cumulative turns reported by Claude.
-    last_reported_turns: tokio::sync::Mutex<u32>,
-}
-
-impl AgentProcess {
-    /// Spawn a persistent Claude process for the named agent.
-    ///
-    /// If a stale session_id causes an immediate exit (e.g. container restart
-    /// cleared the session), automatically retries with a fresh session and
-    /// clears the stale session_id from state. See #149.
-    pub async fn start(name: &str, bus_socket: &str) -> Result<Self> {
-        let state = load_state(name)?;
-        let will_resume =
-            state.config.session == ConfigSessionMode::Persistent && !state.session_id.is_empty();
-
-        let (stdin_tx, event_rx, child) = Self::spawn_process(name, bus_socket, false).await?;
-
-        if will_resume {
-            // Check if the process exits immediately (stale session).
-            // Give it up to 5 seconds — a healthy process stays alive.
-            let mut event_rx_guard = event_rx;
-            let stale_exit = tokio::time::timeout(std::time::Duration::from_secs(5), async {
-                // Peek: if we get ProcessExited quickly, session was stale.
-                loop {
-                    match event_rx_guard.try_recv() {
-                        Ok(StdoutEvent::ProcessExited {
-                            exit_code,
-                            stderr_tail,
-                            ..
-                        }) => return Some((exit_code, stderr_tail)),
-                        Err(tokio::sync::mpsc::error::TryRecvError::Empty) => {
-                            tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-                        }
-                        _ => {
-                            // Got a real event (TextBlock, Result) — process is alive.
-                            return None;
-                        }
-                    }
-                }
-            })
-            .await
-            .unwrap_or(None); // timeout = process is alive
-
-            if let Some((exit_code, stderr_tail)) = stale_exit {
-                let code_str = exit_code
-                    .map(|c| c.to_string())
-                    .unwrap_or_else(|| "?".to_string());
-                warn!(
-                    agent = %name,
-                    session_id = %state.session_id,
-                    exit_code = %code_str,
-                    stderr = %truncate_stderr(&stderr_tail),
-                    "process exited immediately with --resume (stale session), retrying fresh"
-                );
-
-                // Clear stale session_id from state.
-                if let Ok(mut s) = load_state(name) {
-                    s.session_id.clear();
-                    save_state_pub(&s).ok();
-                }
-
-                // Retry with fresh session — verify the retry also stays alive.
-                let (stdin_tx2, event_rx2, child2) =
-                    Self::spawn_process(name, bus_socket, true).await?;
-                let event_rx2 = check_fresh_startup(name, event_rx2).await.map_err(|e| {
-                    error!(
-                        agent = %name,
-                        error = %e,
-                        "stale-session retry also failed — container cannot start"
-                    );
-                    e
-                })?;
-                return Ok(Self {
-                    stdin_tx: stdin_tx2,
-                    event_rx: tokio::sync::Mutex::new(event_rx2),
-                    child: child2,
-                    name: name.to_string(),
-                    last_reported_cost: tokio::sync::Mutex::new(0.0),
-                    last_reported_turns: tokio::sync::Mutex::new(0),
-                });
-            }
-
-            // Process is alive — use the original channels.
-            return Ok(Self {
-                stdin_tx,
-                event_rx: tokio::sync::Mutex::new(event_rx_guard),
-                child,
-                name: name.to_string(),
-                last_reported_cost: tokio::sync::Mutex::new(0.0),
-                last_reported_turns: tokio::sync::Mutex::new(0),
-            });
-        }
-
-        // Fresh start (no --resume): verify the process stays alive before reporting ready.
-        let event_rx = check_fresh_startup(name, event_rx).await?;
-
-        Ok(Self {
-            stdin_tx,
-            event_rx: tokio::sync::Mutex::new(event_rx),
-            child,
-            name: name.to_string(),
-            last_reported_cost: tokio::sync::Mutex::new(0.0),
-            last_reported_turns: tokio::sync::Mutex::new(0),
-        })
-    }
-
-    /// Spawn a persistent Claude process with a fresh session (no --resume).
-    pub async fn start_fresh(name: &str, bus_socket: &str) -> Result<Self> {
-        let (stdin_tx, event_rx, child) = Self::spawn_process(name, bus_socket, true).await?;
-
-        // Verify the process stays alive before reporting it ready.
-        let event_rx = check_fresh_startup(name, event_rx).await?;
-
-        Ok(Self {
-            stdin_tx,
-            event_rx: tokio::sync::Mutex::new(event_rx),
-            child,
-            name: name.to_string(),
-            last_reported_cost: tokio::sync::Mutex::new(0.0),
-            last_reported_turns: tokio::sync::Mutex::new(0),
-        })
-    }
-
-    /// Core spawn logic — builds args, starts child, wires stdin/stdout tasks.
-    /// When  is true, skip --resume regardless of session mode/state.
-    async fn spawn_process(
-        name: &str,
-        bus_socket: &str,
-        fresh: bool,
-    ) -> Result<(
-        tokio::sync::mpsc::UnboundedSender<String>,
-        tokio::sync::mpsc::UnboundedReceiver<StdoutEvent>,
-        std::sync::Arc<tokio::sync::Mutex<Option<tokio::process::Child>>>,
-    )> {
-        let state = load_state(name)?;
-
-        let mut args: Vec<String> = Vec::new();
-
-        // Use --resume only when session is persistent and fresh is not requested.
-        let use_resume = !fresh
-            && state.config.session == ConfigSessionMode::Persistent
-            && !state.session_id.is_empty();
-
-        if use_resume {
-            args.push("--resume".to_string());
-            args.push(state.session_id.clone());
-        }
-
-        if !state.config.system_prompt.is_empty() && !use_resume {
-            args.push("--system-prompt".to_string());
-            args.push(state.config.system_prompt.clone());
-        }
-
-        // Context materialization: load and execute context nodes, inject as system prompt.
-        if let Some(ref ctx_cfg) = state.config.context
-            && ctx_cfg.enabled
-        {
-            let ctx_domain: crate::domain::context::ContextConfig = ctx_cfg.clone().into();
-            let main_path = ctx_cfg
-                .main_path
-                .as_deref()
-                .map(std::path::PathBuf::from)
-                .unwrap_or_else(|| crate::app::context::default_main_path(&state.config.work_dir));
-            if main_path.exists() {
-                match crate::app::context::MainBranch::load(&main_path) {
-                    Ok(mut branch) => {
-                        match branch.materialize().await {
-                            Ok(messages) if !messages.is_empty() => {
-                                let combined: String = messages
-                                    .iter()
-                                    .map(|m| m.content.as_str())
-                                    .collect::<Vec<_>>()
-                                    .join("\n\n");
-                                args.push("--append-system-prompt".to_string());
-                                args.push(combined);
-                                info!(
-                                    agent = %name,
-                                    nodes = messages.len(),
-                                    budget = ctx_domain.main_budget_tokens.unwrap_or(10000),
-                                    "injected materialized context"
-                                );
-                                // Save updated branch (caches live node results).
-                                if let Err(e) = branch.save(&main_path) {
-                                    warn!(agent = %name, error = %e, "failed to save context cache");
-                                }
-                            }
-                            Ok(_) => {
-                                debug!(agent = %name, "context materialized but no messages produced");
-                            }
-                            Err(e) => {
-                                warn!(agent = %name, error = %e, "context materialization failed");
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        warn!(agent = %name, path = %main_path.display(), error = %e, "failed to load context branch");
-                    }
-                }
-            } else {
-                debug!(agent = %name, path = %main_path.display(), "context main path does not exist, skipping");
-            }
-        }
-
-        // Auto-inject required flags for stream-json operation (#151).
-        inject_required_flags(&mut args, &state.config.command);
-
-        // Auto-inject --mcp-config for deskd tools when not already present (#153).
-        // This lets sub-agents use send_message, read_inbox, etc. via MCP.
-        if !state
-            .config
-            .command
-            .iter()
-            .any(|a| a.contains("mcp-config"))
-        {
-            let deskd_bin = std::env::var("DESKD_BIN").unwrap_or_else(|_| "deskd".to_string());
-            let mcp_json = serde_json::json!({
-                "mcpServers": {
-                    "deskd": {
-                        "command": deskd_bin,
-                        "args": ["mcp", "--agent", name]
-                    }
-                }
-            })
-            .to_string();
-            args.push("--mcp-config".to_string());
-            args.push(mcp_json);
-        }
-
-        if !state.config.model.is_empty()
-            && !state
-                .config
-                .command
-                .iter()
-                .any(|a| a == "--model" || a.starts_with("--model="))
-        {
-            args.push("--model".to_string());
-            args.push(state.config.model.clone());
-        }
-
-        let bus_path = bus_socket.to_string();
-        let config_path_str = state.config.config_path.clone().unwrap_or_default();
-        let mut extra_env: Vec<(&str, &str)> =
-            vec![("DESKD_AGENT_NAME", name), ("DESKD_BUS_SOCKET", &bus_path)];
-        if !config_path_str.is_empty() {
-            extra_env.push(("DESKD_AGENT_CONFIG", &config_path_str));
-        }
-
-        let mut cmd = build_command(&state.config, &args, &extra_env);
-        cmd.stdin(Stdio::piped());
-
-        // Log the full command at DEBUG so failures can be reproduced manually.
-        {
-            let std_cmd = cmd.as_std();
-            let program = std_cmd.get_program().to_string_lossy();
-            let cmd_args: Vec<String> = std_cmd
-                .get_args()
-                .map(|a| a.to_string_lossy().into_owned())
-                .collect();
-            debug!(
-                agent = %name,
-                command = %program,
-                args = ?cmd_args,
-                "spawning persistent process"
-            );
-        }
-
-        let mut child = cmd
-            .spawn()
-            .context("Failed to spawn persistent claude process")?;
-
-        info!(agent = %name, model = %state.config.model, "persistent process started");
-
-        let spawn_instant = std::time::Instant::now();
-
-        // Take stdin/stdout.
-        let child_stdin = child.stdin.take().expect("stdin is piped");
-        let stdout = child.stdout.take().expect("stdout is piped");
-        let stderr = child.stderr.take().expect("stderr is piped");
-
-        // Shared stderr buffer (capped at 2 KB) for exit diagnostics.
-        const STDERR_CAP: usize = 2048;
-        let stderr_buf = std::sync::Arc::new(std::sync::Mutex::new(String::new()));
-        let stderr_buf_writer = stderr_buf.clone();
-
-        // Drain stderr in background: persist to file + keep in-memory buffer for exit diagnostics.
-        let agent_name = name.to_string();
-        tokio::spawn(async move {
-            use tokio::io::AsyncBufReadExt;
-
-            let mut reader = tokio::io::BufReader::new(stderr);
-            let log_file = std::fs::OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(stderr_log_path(&agent_name));
-            let mut file_writer = log_file.ok().map(std::io::BufWriter::new);
-
-            let mut line = String::new();
-            while let Ok(n) = reader.read_line(&mut line).await {
-                if n == 0 {
-                    break;
-                }
-                // Persist to file with timestamp.
-                if let Some(ref mut f) = file_writer {
-                    use std::io::Write;
-                    let ts = chrono::Utc::now().format("%H:%M:%S");
-                    let _ = write!(f, "[{}] {}", ts, line);
-                    let _ = f.flush();
-                }
-                // Keep in-memory buffer (capped) for exit diagnostics.
-                if let Ok(mut shared) = stderr_buf_writer.lock() {
-                    shared.push_str(&line);
-                    if shared.len() > STDERR_CAP {
-                        let excess = shared.len() - STDERR_CAP;
-                        shared.drain(..excess);
-                    }
-                }
-                line.clear();
-            }
-        });
-
-        // Stdin writer task.
-        let (stdin_tx, mut stdin_rx) = tokio::sync::mpsc::unbounded_channel::<String>();
-        let mut writer = child_stdin;
-        tokio::spawn(async move {
-            while let Some(line) = stdin_rx.recv().await {
-                if writer.write_all(line.as_bytes()).await.is_err() {
-                    break;
-                }
-            }
-        });
-
-        // Wrap child in Arc so both the reader task and AgentProcess can access it.
-        let child_arc = std::sync::Arc::new(tokio::sync::Mutex::new(Some(child)));
-        let child_for_reader = child_arc.clone();
-
-        // Stdout reader task — parses stream-json and sends events.
-        let (event_tx, event_rx) = tokio::sync::mpsc::unbounded_channel::<StdoutEvent>();
-        let agent_name2 = name.to_string();
-        let stderr_buf_reader = stderr_buf;
-        let task_use_resume = use_resume;
-        tokio::spawn(async move {
-            // Open stream log file for persistence.
-            let stream_log = stream_log_path(&agent_name2);
-            if let Some(parent) = stream_log.parent() {
-                std::fs::create_dir_all(parent).ok();
-            }
-            let stream_file = std::fs::OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(&stream_log);
-            let mut stream_writer = stream_file.ok().map(std::io::BufWriter::new);
-
-            let mut lines = tokio::io::BufReader::new(stdout).lines();
-            while let Ok(Some(line)) = lines.next_line().await {
-                // Persist raw stream-json line.
-                if let Some(ref mut f) = stream_writer {
-                    use std::io::Write;
-                    let _ = writeln!(f, "{}", line);
-                    let _ = f.flush();
-                }
-
-                if let Ok(v) = serde_json::from_str::<serde_json::Value>(&line) {
-                    match v.get("type").and_then(|t| t.as_str()) {
-                        Some("assistant") => {
-                            let mut block_text = String::new();
-                            if let Some(blocks) = v
-                                .get("message")
-                                .and_then(|m| m.get("content"))
-                                .and_then(|c| c.as_array())
-                            {
-                                for block in blocks {
-                                    if block.get("type").and_then(|t| t.as_str()) == Some("text")
-                                        && let Some(text) =
-                                            block.get("text").and_then(|t| t.as_str())
-                                    {
-                                        block_text.push_str(text);
-                                    }
-                                }
-                            }
-                            // Extract token usage from assistant message.
-                            let usage = v.get("message").and_then(|m| m.get("usage")).map(|u| {
-                                let mut tu = TokenUsage::default();
-                                tu.accumulate(u);
-                                tu
-                            });
-                            if (!block_text.is_empty() || usage.is_some())
-                                && event_tx
-                                    .send(StdoutEvent::TextBlock(block_text, usage))
-                                    .is_err()
-                            {
-                                break;
-                            }
-                        }
-                        Some("result") => {
-                            let session_id = v
-                                .get("session_id")
-                                .and_then(|s| s.as_str())
-                                .unwrap_or_default()
-                                .to_string();
-                            let cost_usd = v
-                                .get("total_cost_usd")
-                                .and_then(|c| c.as_f64())
-                                .unwrap_or(0.0);
-                            let num_turns =
-                                v.get("num_turns").and_then(|t| t.as_u64()).unwrap_or(0) as u32;
-                            // response_text and token_usage are accumulated by send_task, not here.
-                            if event_tx
-                                .send(StdoutEvent::Result(TurnResult {
-                                    response_text: String::new(),
-                                    session_id,
-                                    cost_usd,
-                                    num_turns,
-                                    token_usage: TokenUsage::default(),
-                                }))
-                                .is_err()
-                            {
-                                break;
-                            }
-                            // Rotate stream log after each turn if over 10MB.
-                            rotate_stream_log(&stream_log, 10 * 1024 * 1024);
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            // stdout closed — process exited. Capture exit context.
-            let lifetime_secs = spawn_instant.elapsed().as_secs();
-            let exit_code = if let Some(mut ch) = child_for_reader.lock().await.take() {
-                ch.wait().await.ok().and_then(|s| s.code())
-            } else {
-                None // already taken by kill()
-            };
-            // Give stderr task a moment to finish writing.
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-            let stderr_tail = stderr_buf_reader
-                .lock()
-                .map(|s| s.clone())
-                .unwrap_or_default();
-
-            let _ = event_tx.send(StdoutEvent::ProcessExited {
-                exit_code,
-                stderr_tail,
-                lifetime_secs,
-                used_resume: task_use_resume,
-            });
-            debug!(agent = %agent_name2, "persistent process stdout closed");
-        });
-
-        Ok((stdin_tx, event_rx, child_arc))
-    }
-
-    /// Send a task to the persistent process and collect the response.
-    ///
-    /// Enforces `limits` in real-time: if max_turns or budget is exceeded
-    /// mid-task, the process is killed immediately.
-    pub async fn send_task(
-        &self,
-        message: &str,
-        progress_tx: Option<&tokio::sync::mpsc::UnboundedSender<String>>,
-        image: Option<(&str, &str)>,
-        limits: &TaskLimits,
-    ) -> Result<TurnResult> {
-        // Drain any stale events left over from the previous turn before
-        // starting a new one. The post-result drain (below) catches most
-        // trailing TextBlock events, but there is a race window: if the
-        // stdout reader task has not yet queued events when try_recv()
-        // runs after the result, those events survive into the next turn.
-        // Draining here, before writing to stdin, ensures each turn
-        // starts with a clean channel. (#102)
-        {
-            let mut event_rx = self.event_rx.lock().await;
-            let mut drained = 0u32;
-            loop {
-                match event_rx.try_recv() {
-                    Ok(StdoutEvent::ProcessExited {
-                        exit_code,
-                        stderr_tail,
-                        lifetime_secs,
-                        used_resume,
-                    }) => {
-                        bail!(
-                            "persistent process exited between turns \
-                             (exit_code={}, lifetime={}s, resume={}, stderr={:?})",
-                            exit_code
-                                .map(|c| c.to_string())
-                                .unwrap_or_else(|| "signal".into()),
-                            lifetime_secs,
-                            used_resume,
-                            truncate_stderr(&stderr_tail),
-                        );
-                    }
-                    Ok(_) => {
-                        drained += 1;
-                    }
-                    Err(_) => break,
-                }
-            }
-            if drained > 0 {
-                info!(
-                    agent = %self.name,
-                    drained_events = drained,
-                    "flushed stale stdout events before new turn"
-                );
-            }
-            drop(event_rx);
-            // Yield to let the stdout reader task push any in-flight
-            // data before we proceed.
-            tokio::task::yield_now().await;
-            // Second drain pass after yielding — catches events that were
-            // being processed during the first pass.
-            let mut event_rx = self.event_rx.lock().await;
-            loop {
-                match event_rx.try_recv() {
-                    Ok(StdoutEvent::ProcessExited {
-                        exit_code,
-                        stderr_tail,
-                        lifetime_secs,
-                        used_resume,
-                    }) => {
-                        bail!(
-                            "persistent process exited between turns \
-                             (exit_code={}, lifetime={}s, resume={}, stderr={:?})",
-                            exit_code
-                                .map(|c| c.to_string())
-                                .unwrap_or_else(|| "signal".into()),
-                            lifetime_secs,
-                            used_resume,
-                            truncate_stderr(&stderr_tail),
-                        );
-                    }
-                    Ok(_) => {
-                        drained += 1;
-                    }
-                    Err(_) => break,
-                }
-            }
-            if drained > 0 {
-                info!(
-                    agent = %self.name,
-                    total_drained = drained,
-                    "pre-turn drain complete"
-                );
-            }
-        }
-
-        // Build the user message.
-        let user_msg = if let Some((b64_data, media_type)) = image {
-            let msg = serde_json::json!({
-                "type": "user",
-                "message": {
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "image",
-                            "source": {
-                                "type": "base64",
-                                "media_type": media_type,
-                                "data": b64_data,
-                            }
-                        },
-                        {"type": "text", "text": message}
-                    ]
-                }
-            });
-            let mut line = serde_json::to_string(&msg)?;
-            line.push('\n');
-            line
-        } else {
-            format_user_message(message)
-        };
-
-        // Send to stdin.
-        self.stdin_tx
-            .send(user_msg)
-            .map_err(|_| anyhow::anyhow!("persistent process stdin closed"))?;
-
-        // Read events until we get a Result, enforcing limits on each event.
-        let mut event_rx = self.event_rx.lock().await;
-        let mut response_text = String::new();
-        let mut assistant_turns = 0u32;
-        let mut accumulated_usage = TokenUsage::default();
-
-        loop {
-            match event_rx.recv().await {
-                Some(StdoutEvent::TextBlock(text, usage)) => {
-                    // Accumulate token usage from this assistant message.
-                    if let Some(u) = &usage {
-                        accumulated_usage.merge(u);
-                    }
-
-                    if !text.is_empty() {
-                        // Only count turns with actual text content, not
-                        // tool-use messages that only carry usage metadata.
-                        assistant_turns += 1;
-
-                        // Check turn limit.
-                        if let Some(max) = limits.max_turns
-                            && assistant_turns > max
-                        {
-                            warn!(
-                                agent = %self.name,
-                                turns = assistant_turns,
-                                max = max,
-                                "turn limit exceeded mid-task, killing process"
-                            );
-                            self.kill().await;
-                            bail!(
-                                "task killed: exceeded {} turn limit ({} turns)",
-                                max,
-                                assistant_turns
-                            );
-                        }
-
-                        if let Some(tx) = &progress_tx {
-                            let _ = tx.send(text.clone());
-                        }
-                        response_text.push_str(&text);
-                    }
-                }
-                Some(StdoutEvent::Result(mut result)) => {
-                    // Drain any trailing TextBlock events that arrived after the
-                    // result event but belong to this same turn. Without this,
-                    // the next call to send_task() would see stale blocks from
-                    // the previous turn, causing an off-by-one where turn N's
-                    // output appears as part of turn N+1's response. (#102)
-                    while let Ok(StdoutEvent::TextBlock(trailing, _)) = event_rx.try_recv() {
-                        debug!(
-                            agent = %self.name,
-                            len = trailing.len(),
-                            "drained trailing text block after result event"
-                        );
-                        if let Some(tx) = &progress_tx {
-                            let _ = tx.send(trailing.clone());
-                        }
-                        response_text.push_str(&trailing);
-                    }
-
-                    result.response_text = response_text;
-                    result.token_usage = accumulated_usage;
-
-                    // Compute deltas: Claude's total_cost_usd and num_turns are
-                    // session-cumulative, not per-task. We track the last reported
-                    // values to compute the actual delta for this task.
-                    let mut last_cost = self.last_reported_cost.lock().await;
-                    let mut last_turns = self.last_reported_turns.lock().await;
-                    let cost_delta = (result.cost_usd - *last_cost).max(0.0);
-                    let turns_delta = result.num_turns.saturating_sub(*last_turns);
-                    *last_cost = result.cost_usd;
-                    *last_turns = result.num_turns;
-                    drop(last_cost);
-                    drop(last_turns);
-
-                    // Update state file with deltas.
-                    if let Ok(mut state) = load_state(&self.name) {
-                        // Only save session_id for persistent agents.
-                        if state.config.session == ConfigSessionMode::Persistent
-                            && !result.session_id.is_empty()
-                        {
-                            state.session_id = result.session_id.clone();
-                        }
-                        state.total_cost += cost_delta;
-                        state.total_turns += turns_delta;
-                        let _ = save_state(&state);
-
-                        // Check budget limit after updating cost.
-                        if let Some(budget) = limits.budget_usd
-                            && state.total_cost >= budget
-                        {
-                            warn!(
-                                agent = %self.name,
-                                cost = state.total_cost,
-                                budget = budget,
-                                "budget exceeded, killing process"
-                            );
-                            self.kill().await;
-                        }
-                    }
-
-                    return Ok(result);
-                }
-                Some(StdoutEvent::ProcessExited {
-                    exit_code,
-                    stderr_tail,
-                    lifetime_secs,
-                    used_resume,
-                }) => {
-                    bail!(
-                        "persistent process exited mid-task \
-                         (exit_code={}, lifetime={}s, resume={}, stderr={:?})",
-                        exit_code
-                            .map(|c| c.to_string())
-                            .unwrap_or_else(|| "signal".into()),
-                        lifetime_secs,
-                        used_resume,
-                        truncate_stderr(&stderr_tail),
-                    );
-                }
-                None => {
-                    bail!("persistent process exited mid-task (channel closed)");
-                }
-            }
-        }
-    }
-
-    /// Inject a message into the running process as a new user turn.
-    pub fn inject_message(&self, message: &str) -> Result<()> {
-        let line = format_user_message(message);
-        self.stdin_tx
-            .send(line)
-            .map_err(|_| anyhow::anyhow!("persistent process stdin closed"))
-    }
-
-    /// Kill the running process immediately (e.g. budget exceeded mid-task).
-    /// The process can be restarted later with a fresh AgentProcess::start().
-    pub async fn kill(&self) {
-        if let Some(mut child) = self.child.lock().await.take() {
-            let _ = child.kill().await;
-            let _ = child.wait().await;
-        }
-        warn!(agent = %self.name, "persistent process killed");
-    }
-
-    /// Gracefully stop the persistent process.
-    pub async fn stop(&self) {
-        // Dropping all senders closes stdin, which causes Claude to exit.
-        // We can't drop self.stdin_tx (owned), but we can kill the child.
-        if let Some(mut child) = self.child.lock().await.take() {
-            let _ = child.kill().await;
-            let _ = child.wait().await;
-        }
-        info!(agent = %self.name, "persistent process stopped");
-    }
-}
-
-impl Executor for AgentProcess {
-    fn send_task<'a>(
-        &'a self,
-        message: &'a str,
-        progress_tx: Option<&'a tokio::sync::mpsc::UnboundedSender<String>>,
-        image: Option<(&'a str, &'a str)>,
-        limits: &'a TaskLimits,
-    ) -> std::pin::Pin<Box<dyn std::future::Future<Output = Result<TurnResult>> + Send + 'a>> {
-        Box::pin(self.send_task(message, progress_tx, image, limits))
-    }
-
-    fn inject_message(&self, message: &str) -> Result<()> {
-        self.inject_message(message)
-    }
-
-    fn stop(&self) -> std::pin::Pin<Box<dyn std::future::Future<Output = ()> + Send + '_>> {
-        Box::pin(self.stop())
-    }
-}
-
+// Tests for TokenUsage live in ports::executor. Only keep re-export-level tests here.
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_agent_config_defaults() {
-        let yaml = r#"
-config:
-  name: test
-  model: claude-opus-4-6
-  system_prompt: ""
-  work_dir: /tmp
-  max_turns: 100
-pid: 0
-session_id: ""
-total_turns: 0
-total_cost: 0.0
-created_at: "2024-01-01T00:00:00Z"
-"#;
-        let state: AgentState = serde_yaml::from_str(yaml).unwrap();
-        assert_eq!(state.config.budget_usd, 50.0);
-        assert!(state.config.unix_user.is_none());
-    }
-
-    #[test]
-    fn test_agent_state_round_trip() {
-        let cfg = AgentConfig {
-            name: "test-agent".to_string(),
-            model: "claude-sonnet-4-6".to_string(),
-            system_prompt: String::new(),
-            work_dir: "/tmp".to_string(),
-            max_turns: 100,
-            unix_user: Some("agent1".to_string()),
-            budget_usd: 10.0,
-            command: vec!["claude".to_string()],
-            config_path: Some("/home/agent1/deskd.yaml".to_string()),
-            container: None,
-            session: ConfigSessionMode::default(),
-            runtime: ConfigAgentRuntime::default(),
-            context: None,
-        };
-        let state = AgentState {
-            config: cfg,
-            pid: 0,
-            session_id: "sid-123".to_string(),
-            total_turns: 5,
-            total_cost: 0.42,
-            created_at: Utc::now().to_rfc3339(),
-            status: "idle".to_string(),
-            current_task: String::new(),
-            parent: None,
-        };
-        let yaml = serde_yaml::to_string(&state).unwrap();
-        let restored: AgentState = serde_yaml::from_str(&yaml).unwrap();
-        assert_eq!(restored.config.unix_user.as_deref(), Some("agent1"));
-        assert_eq!(restored.session_id, "sid-123");
-        assert_eq!(restored.total_cost, 0.42);
-        assert_eq!(
-            restored.config.config_path.as_deref(),
-            Some("/home/agent1/deskd.yaml")
-        );
-    }
-
-    #[test]
-    fn test_format_user_message() {
-        let line = format_user_message("hello world");
-        let v: serde_json::Value = serde_json::from_str(line.trim()).unwrap();
-        assert_eq!(v["type"], "user");
-        assert_eq!(v["message"]["role"], "user");
-        let content = v["message"]["content"].as_array().unwrap();
-        assert_eq!(content.len(), 1);
-        assert_eq!(content[0]["type"], "text");
-        assert_eq!(content[0]["text"], "hello world");
-    }
-
-    #[test]
-    fn test_expand_tilde_simple() {
-        let result = expand_tilde("/absolute/path");
-        assert_eq!(result, "/absolute/path");
-    }
-
-    #[test]
-    fn test_expand_tilde_home() {
-        let home = std::env::var("HOME").unwrap();
-        let result = expand_tilde("~/.ssh");
-        assert_eq!(result, format!("{}/.ssh", home));
-    }
-
-    #[test]
-    fn test_expand_tilde_with_ro() {
-        let home = std::env::var("HOME").unwrap();
-        // expand_tilde treats "ro" as a path segment; normalize_mount fixes it later.
-        let result = expand_tilde("~/.ssh:ro");
-        assert_eq!(result, format!("{}/.ssh:ro", home));
-    }
-
-    #[test]
-    fn test_normalize_mount_path_only() {
-        assert_eq!(normalize_mount("/foo"), "/foo:/foo");
-    }
-
-    #[test]
-    fn test_normalize_mount_path_ro() {
-        assert_eq!(normalize_mount("/foo:ro"), "/foo:/foo:ro");
-    }
-
-    #[test]
-    fn test_normalize_mount_src_dst() {
-        assert_eq!(normalize_mount("/foo:/bar"), "/foo:/bar");
-    }
-
-    #[test]
-    fn test_normalize_mount_src_dst_ro() {
-        assert_eq!(normalize_mount("/foo:/bar:ro"), "/foo:/bar:ro");
-    }
-
-    #[test]
-    fn test_expand_tilde_src_dst() {
-        let home = std::env::var("HOME").unwrap();
-        let result = expand_tilde("~/.gitconfig:~/.gitconfig:ro");
-        assert_eq!(
-            result,
-            format!("{}/.gitconfig:{}/.gitconfig:ro", home, home)
-        );
-    }
-
-    #[test]
-    fn test_build_container_command() {
-        use std::collections::HashMap;
-        let mut env = HashMap::new();
-        env.insert("GH_TOKEN".to_string(), "test-token".to_string());
-
-        let container = ContainerConfig {
-            image: "claude-code-local:official".to_string(),
-            mounts: vec!["/host/path:/container/path:ro".to_string()],
-            volumes: vec!["my-vol:/data".to_string()],
-            env,
-            runtime: "docker".to_string(),
-        };
-
-        let cfg = AgentConfig {
-            name: "test-agent".to_string(),
-            model: "claude-sonnet-4-6".to_string(),
-            system_prompt: String::new(),
-            work_dir: "/home/test".to_string(),
-            max_turns: 100,
-            unix_user: None,
-            budget_usd: 50.0,
-            command: vec![
-                "claude".to_string(),
-                "--output-format".to_string(),
-                "stream-json".to_string(),
-            ],
-            config_path: Some("/home/test/deskd.yaml".to_string()),
-            container: Some(container),
-            session: ConfigSessionMode::default(),
-            runtime: ConfigAgentRuntime::default(),
-            context: None,
-        };
-
-        let extra_env = [("DESKD_BUS_SOCKET", "/home/test/.deskd/bus.sock")];
-        let args = vec!["--resume".to_string(), "session-1".to_string()];
-
-        let cmd = build_command(&cfg, &args, &extra_env);
-        let program = cmd.as_std().get_program().to_string_lossy().to_string();
-        assert_eq!(program, "docker");
-
-        let cmd_args: Vec<String> = cmd
-            .as_std()
-            .get_args()
-            .map(|a| a.to_string_lossy().to_string())
-            .collect();
-
-        // Verify basic structure.
-        assert!(cmd_args.contains(&"run".to_string()));
-        assert!(cmd_args.contains(&"--rm".to_string()));
-        assert!(cmd_args.contains(&"-i".to_string()));
-        assert!(cmd_args.contains(&"deskd-test-agent".to_string()));
-        assert!(cmd_args.contains(&"/home/test:/home/test".to_string()));
-        assert!(cmd_args.contains(&"/home/test".to_string()));
-        assert!(cmd_args.contains(&"claude-code-local:official".to_string()));
-        assert!(cmd_args.contains(&"/host/path:/container/path:ro".to_string()));
-        assert!(cmd_args.contains(&"my-vol:/data".to_string()));
-        assert!(cmd_args.contains(&"GH_TOKEN=test-token".to_string()));
-        assert!(cmd_args.contains(&"DESKD_BUS_SOCKET=/home/test/.deskd/bus.sock".to_string()));
-        // The agent command + args should be at the end.
-        assert!(cmd_args.contains(&"claude".to_string()));
-        assert!(cmd_args.contains(&"--output-format".to_string()));
-        assert!(cmd_args.contains(&"stream-json".to_string()));
-        assert!(cmd_args.contains(&"--resume".to_string()));
-        assert!(cmd_args.contains(&"session-1".to_string()));
-    }
-
-    #[test]
-    fn test_build_command_no_container() {
-        let cfg = AgentConfig {
-            name: "test".to_string(),
-            model: "claude-sonnet-4-6".to_string(),
-            system_prompt: String::new(),
-            work_dir: "/tmp".to_string(),
-            max_turns: 100,
-            unix_user: None,
-            budget_usd: 50.0,
-            command: vec!["claude".to_string()],
-            config_path: None,
-            container: None,
-            session: ConfigSessionMode::default(),
-            runtime: ConfigAgentRuntime::default(),
-            context: None,
-        };
-        let cmd = build_command(&cfg, &[], &[]);
-        let program = cmd.as_std().get_program().to_string_lossy().to_string();
-        assert_eq!(program, "claude");
-    }
 
     #[test]
     fn test_token_usage_default() {
@@ -1960,7 +47,6 @@ created_at: "2024-01-01T00:00:00Z"
         assert_eq!(usage.cache_creation_input_tokens, 10);
         assert_eq!(usage.cache_read_input_tokens, 20);
 
-        // Accumulate a second message.
         let json2 = serde_json::json!({
             "input_tokens": 200,
             "output_tokens": 100,
@@ -1978,7 +64,6 @@ created_at: "2024-01-01T00:00:00Z"
     fn test_token_usage_accumulate_partial() {
         let mut usage = TokenUsage::default();
 
-        // Only some fields present.
         let json = serde_json::json!({
             "input_tokens": 100,
             "output_tokens": 50
@@ -2009,144 +94,5 @@ created_at: "2024-01-01T00:00:00Z"
         assert_eq!(a.output_tokens, 125);
         assert_eq!(a.cache_creation_input_tokens, 15);
         assert_eq!(a.cache_read_input_tokens, 35);
-    }
-
-    #[test]
-    fn test_split_command_empty() {
-        let cmd: Vec<String> = vec![];
-        let (bin, args) = split_command(&cmd);
-        assert_eq!(bin, "claude");
-        assert!(args.is_empty());
-    }
-
-    #[test]
-    fn test_split_command_single() {
-        let cmd = vec!["my-agent".to_string()];
-        let (bin, args) = split_command(&cmd);
-        assert_eq!(bin, "my-agent");
-        assert!(args.is_empty());
-    }
-
-    #[test]
-    fn test_split_command_with_args() {
-        let cmd = vec![
-            "claude".to_string(),
-            "--output-format".to_string(),
-            "stream-json".to_string(),
-        ];
-        let (bin, args) = split_command(&cmd);
-        assert_eq!(bin, "claude");
-        assert_eq!(args.len(), 2);
-        assert_eq!(args[0], "--output-format");
-    }
-
-    #[test]
-    fn test_inject_required_flags_empty_command() {
-        let mut args = Vec::new();
-        let command = vec!["claude".to_string()];
-        inject_required_flags(&mut args, &command);
-        assert!(args.contains(&"--input-format=stream-json".to_string()));
-        assert!(args.contains(&"--output-format".to_string()));
-        assert!(args.contains(&"--verbose".to_string()));
-        assert!(args.contains(&"--dangerously-skip-permissions".to_string()));
-    }
-
-    #[test]
-    fn test_inject_required_flags_no_duplicates() {
-        let mut args = Vec::new();
-        let command = vec![
-            "claude".to_string(),
-            "--output-format".to_string(),
-            "stream-json".to_string(),
-            "--verbose".to_string(),
-            "--dangerously-skip-permissions".to_string(),
-        ];
-        inject_required_flags(&mut args, &command);
-        // Only input-format should be injected; others already in command.
-        assert!(args.contains(&"--input-format=stream-json".to_string()));
-        assert!(!args.contains(&"--verbose".to_string()));
-        assert!(!args.contains(&"--dangerously-skip-permissions".to_string()));
-        // output-format should NOT be injected since command already has it.
-        let output_format_count = args.iter().filter(|a| a.contains("output-format")).count();
-        assert_eq!(output_format_count, 0);
-    }
-
-    #[test]
-    fn test_state_save_load_roundtrip() {
-        let tmp =
-            std::env::temp_dir().join(format!("deskd-test-agent-state-{}", uuid::Uuid::new_v4()));
-        std::fs::create_dir_all(&tmp).unwrap();
-        unsafe { std::env::set_var("HOME", &tmp) };
-
-        let cfg = AgentConfig {
-            name: "roundtrip-test".to_string(),
-            model: "claude-sonnet-4-6".to_string(),
-            system_prompt: "test prompt".to_string(),
-            work_dir: "/tmp".to_string(),
-            max_turns: 50,
-            unix_user: None,
-            budget_usd: 25.0,
-            command: vec!["claude".to_string()],
-            config_path: None,
-            container: None,
-            session: ConfigSessionMode::default(),
-            runtime: ConfigAgentRuntime::default(),
-            context: None,
-        };
-        let state = AgentState {
-            config: cfg,
-            pid: 0,
-            session_id: "sess-xyz".to_string(),
-            total_turns: 10,
-            total_cost: 1.23,
-            created_at: "2026-01-01T00:00:00Z".to_string(),
-            status: "idle".to_string(),
-            current_task: String::new(),
-            parent: Some("parent-agent".to_string()),
-        };
-
-        save_state(&state).unwrap();
-        let loaded = load_state("roundtrip-test").unwrap();
-
-        assert_eq!(loaded.config.name, "roundtrip-test");
-        assert_eq!(loaded.session_id, "sess-xyz");
-        assert_eq!(loaded.total_turns, 10);
-        assert_eq!(loaded.total_cost, 1.23);
-        assert_eq!(loaded.parent.as_deref(), Some("parent-agent"));
-        assert_eq!(loaded.config.budget_usd, 25.0);
-
-        let _ = std::fs::remove_dir_all(&tmp);
-    }
-
-    #[test]
-    fn test_build_command_with_unix_user() {
-        let cfg = AgentConfig {
-            name: "test".to_string(),
-            model: "claude-sonnet-4-6".to_string(),
-            system_prompt: String::new(),
-            work_dir: "/tmp".to_string(),
-            max_turns: 100,
-            unix_user: Some("agent-user".to_string()),
-            budget_usd: 50.0,
-            command: vec!["claude".to_string()],
-            config_path: None,
-            container: None,
-            session: ConfigSessionMode::default(),
-            runtime: ConfigAgentRuntime::default(),
-            context: None,
-        };
-        let extra_env = [("DESKD_BUS_SOCKET", "/tmp/bus.sock")];
-        let cmd = build_command(&cfg, &[], &extra_env);
-        let program = cmd.as_std().get_program().to_string_lossy().to_string();
-        assert_eq!(program, "sudo");
-
-        let args: Vec<String> = cmd
-            .as_std()
-            .get_args()
-            .map(|a| a.to_string_lossy().to_string())
-            .collect();
-        assert!(args.contains(&"-u".to_string()));
-        assert!(args.contains(&"agent-user".to_string()));
-        assert!(args.contains(&"DESKD_BUS_SOCKET=/tmp/bus.sock".to_string()));
     }
 }
